Provide an example database.yml and use that.
This commit moves the provided database.yml to database.example.yml so
there is a documented config file that can just be copied into place,
but if you have one in place, you don't have to edit it when pushing
commits back up. This also places database.yml in the gitignore list
since it's local configuration and not part of the software package.